polish wrapper MissingNodeIdToStatusMap
